+++ artifacts/dev/tac/structs.ts
@@ -1,1 +1,1 @@
-export type OutMessageV1Struct = import('../l2-evm/typechain-types/contracts/core/Structs.sol/IStructsInterface').OutMessageV1Struct;
+export type OutMessageV1Struct = import('./internal/typechain-types/contracts/core/Structs.sol/IStructsInterface').OutMessageV1Struct;