+++ artifacts/testnet/tac/utils.ts
@@ -1,1 +1,1 @@
-export { encodeOutMessageV1 } from '../l2-evm/scripts/utils/encoding';
+export { encodeOutMessageV1 } from './internal/scripts/utils/encoding';